Updated Recycler view with course info.

+++ app/src/main/java/com/example/cst438_project01_group4/ClassObjects/Grade.java
@@ -2,21 +2,22 @@ package com.example.cst438_project01_group4.ClassObjects;
 
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
-import java.util.Date;
 
 import com.example.cst438_project01_group4.DataBase.AppDatabase;
+
+import java.util.Date;
 
 @Entity(tableName = AppDatabase.GRADE_TABLE)
 public class Grade {
     @PrimaryKey(autoGenerate = true)
     private int gradeID;
 
-    private int score;
+    private double score;
     private int assignmentID;
     private int courseID;
     private Date dateEarned;
 
-    public Grade(int score, int assignmentID, int courseID, Date dateEarned) {
+    public Grade(double score, int assignmentID, int courseID, Date dateEarned) {
         this.score = score;
         this.assignmentID = assignmentID;
         this.courseID = courseID;
@@ -31,11 +32,11 @@ public class Grade {
         this.gradeID = gradeID;
     }
 
-    public int getScore() {
+    public double getScore() {
         return score;
     }
 
-    public void setScore(int score) {
+    public void setScore(double score) {
         this.score = score;
     }
